Created RoomDTO

+++ src/main/java/com/codehub/theater_management/repository/TheaterRepository.java
@@ -3,5 +3,5 @@ package com.codehub.theater_management.repository;
 import com.codehub.theater_management.model.Theater;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TheaterRepository extends JpaRepository<Theater, Integer> {
+public interface TheaterRepository extends JpaRepository<Theater, Long> {
 }
